+++ Modules/Intent.Modules.Modelers.UI.Core/Api/RadioGroupModel.cs
@@ -1,0 +1,80 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Intent.Metadata.Models;
+using Intent.Modules.Common;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.Api.ApiElementModel", Version = "1.0")]
+
+namespace Intent.Modelers.UI.Core.Api
+{
+    [IntentManaged(Mode.Fully, Signature = Mode.Fully)]
+    public class RadioGroupModel : IMetadataModel, IHasStereotypes, IHasName, IElementWrapper
+    {
+        public const string SpecializationType = "Radio Group";
+        public const string SpecializationTypeId = "4af9a7f0-ca24-4681-8e32-2a8d3dacb6ed";
+        protected readonly IElement _element;
+
+        [IntentManaged(Mode.Fully)]
+        public RadioGroupModel(IElement element, string requiredType = SpecializationType)
+        {
+            if (!requiredType.Equals(element.SpecializationType, StringComparison.InvariantCultureIgnoreCase))
+            {
+                throw new Exception($"Cannot create a '{GetType().Name}' from element with specialization type '{element.SpecializationType}'. Must be of type '{SpecializationType}'");
+            }
+            _element = element;
+        }
+
+        public string Id => _element.Id;
+
+        public string Name => _element.Name;
+
+        public string Comment => _element.Comment;
+
+        public IEnumerable<IStereotype> Stereotypes => _element.Stereotypes;
+
+        public string Value => _element.Value;
+
+        public IElement InternalElement => _element;
+
+        public override string ToString()
+        {
+            return _element.ToString();
+        }
+
+        public bool Equals(RadioGroupModel other)
+        {
+            return Equals(_element, other?._element);
+        }
+
+        public override bool Equals(object obj)
+        {
+            if (ReferenceEquals(null, obj)) return false;
+            if (ReferenceEquals(this, obj)) return true;
+            if (obj.GetType() != this.GetType()) return false;
+            return Equals((RadioGroupModel)obj);
+        }
+
+        public override int GetHashCode()
+        {
+            return (_element != null ? _element.GetHashCode() : 0);
+        }
+    }
+
+    [IntentManaged(Mode.Fully)]
+    public static class RadioGroupModelExtensions
+    {
+
+        public static bool IsRadioGroupModel(this ICanBeReferencedType type)
+        {
+            return type != null && type is IElement element && element.SpecializationTypeId == RadioGroupModel.SpecializationTypeId;
+        }
+
+        public static RadioGroupModel AsRadioGroupModel(this ICanBeReferencedType type)
+        {
+            return type.IsRadioGroupModel() ? new RadioGroupModel((IElement)type) : null;
+        }
+    }
+}