+++ Modules/Intent.Modules.Modelers.UI.Core/release-notes.md
@@ -3,3 +3,4 @@
 - Initial release.
 - Improvement : Support for mapping `Menu Item`s which are children of `Menu Item`s.
 - Improvement : Add `On Selected` event for `Select` components.
+- Improvement : Support for `Radio Button Group`